Use ints as for loop interators.
Resolves compiler warnings about converting size_t to int, e.g. VS
warning C4267.

+++ source/src/ports/generic_networkhandler.c
@@ -447,7 +447,7 @@ EipStatus NetworkHandlerProcessOnce(void) {
     CheckAndHandleUdpGlobalBroadcastSocket();
     CheckAndHandleConsumingUdpSockets();
 
-    for(size_t socket = 0; socket <= highest_socket_handle; socket++) {
+    for(int socket = 0; socket <= highest_socket_handle; socket++) {
       if(true == CheckSocketSet(socket) ) {
         /* if it is still checked it is a TCP receive */
         if(kEipStatusError == HandleDataOnTcpSocket(socket) ) /* if error */
@@ -459,7 +459,7 @@ EipStatus NetworkHandlerProcessOnce(void) {
     }
   }
 
-  for(size_t socket = 0; socket <= highest_socket_handle; socket++) {
+  for(int socket = 0; socket <= highest_socket_handle; socket++) {
     CheckEncapsulationInactivity(socket);
   }
